requirements.txt
appdirs==1.4.3
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
colorama==0.4.3
contextlib2==0.6.0
cycler==0.10.0
dataclasses==0.6
distlib==0.3.0
distro==1.4.0
django-environ==0.4.5
Flask==1.1.2
future==0.18.2
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.3.1
lockfile==0.12.2
MarkupSafe==1.1.1
matplotlib==3.3.3
msgpack==0.6.2
numpy==1.19.4
opencv-python==4.4.0.46
packaging==20.3
pep517==0.8.2
Pillow==8.0.1
progress==1.5
pyparsing==2.4.6
python-dateutil==2.8.1
pytoml==0.1.21
requests==2.22.0
retrying==1.3.3
six==1.14.0
torch==1.7.0
torchvision==0.8.1
tqdm==4.54.1
typing-extensions==3.7.4.3
urllib3==1.25.8
webencodings==0.5.1
Werkzeug==1.0.1
